tests/wpt/tests/css/css-masking/animations/clip-path-interpolation-shape.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>clip-path-interpolation</title>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-clip-path">
<link rel="help" href="https://drafts.csswg.org/css-shapes-2/#interpolating-shape">
<meta name="assert" content="clip-path supports animation for shape()">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
html {
  font-size: 16px;
  font-family: Ahem;
}

.parent {
  clip-path: shape(from -5px 5px, move to 5% 1px);
  padding: 10px;
}

.target {
  display: inline-block;
  width: 100px;
  height: 100px;
  background-color: black;
  clip-path: shape(from 5px 5px, line to 10px 10%);
}

.expected {
  background-color: green;
}
</style>
<body>
<script>
test_interpolation({
  property: 'clip-path',
  from: neutralKeyframe,
  to: 'shape(from -5px 5px, line to 20px 20%)',
}, [
  {at: -0.3, expect: 'shape(from 8px 5px, line to 7px 7%)'},
  {at: 0, expect: 'shape(from 5px 5px, line to 10px 10%)'},
  {at: 0.6, expect: 'shape(from -1px 5px, line to 16px 16%)'},
  {at: 1, expect: 'shape(from -5px 5px, line to 20px 20%)'},
  {at: 1.5, expect: 'shape(from -10px 5px, line to 25px 25%)'},
]);

test_no_interpolation({
  property: 'clip-path',
  from: 'initial',
  to: 'shape(from 8px 5px, line to 7px 7%)',
});

test_interpolation({
  property: 'clip-path',
  from: 'inherit',
  to: 'shape(from 15% 15px, move to 20% -10px)',
}, [
  {at: -0.3, expect: 'shape(from calc(-4.5% - 6.5px) 2px, move to 0.5% 4.3px)'},
  {at: 0, expect: 'shape(from calc(0% - 5px) 5px, move to 5% 1px)'},
  {at: 0.5, expect: 'shape(from calc(7.5% - 2.5px) 10px, move to 12.5% -4.5px)'},
  {at: 1, expect: 'shape(from 15% 15px, move to 20% -10px'},
  {at: 1.5, expect: 'shape(from calc(22.5% + 2.5px) 20px, move to 27.5% -15.5px)'},
]);

test_no_interpolation({
  property: 'clip-path',
  from: 'unset',
  to: 'shape(from 10px 10px, close)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'none',
  to: 'shape(from 10px 10px, close)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'shape(from 10px 10px, move to 10% 10%)',
  to: 'shape(from 10px 10px, close)',
});

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, hline to 5%, vline to -5px, close)',
  to: 'shape(from 15% 15px, hline to 25%, vline to -15px, close)',
}, [
  {at: -0.3, expect: 'shape(from 2% 2px, hline to -1%, vline to -2px, close)'},
  {at: 0, expect: 'shape(from 5% 5px, hline to 5%, vline to -5px, close)'},
  {at: 0.5, expect: 'shape(from 10% 10px, hline to 15% , vline to -10px, close)'},
  {at: 1, expect: 'shape(from 15% 15px, hline to 25%, vline to -15px, close)'},
  {at: 1.5, expect: 'shape(from 20% 20px, hline to 35%, vline to -20px, close)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, curve to 10% 10px with 0% 80px, curve to 30% 20px with 20% 50px / 25% 70px)',
  to: 'shape(from 15% 15px, curve to 20% 0px with 10% 60px, curve to 20% 30px with 30% 40px / -5% 100px)',
}, [
  {at: -0.3, expect: 'shape(from 2% 2px, curve to 7% 13px with -3% 86px, curve to 33% 17px with 17% 53px / 34% 61px)'},
  {at: 0, expect: 'shape(from 5% 5px, curve to 10% 10px with 0% 80px, curve to 30% 20px with 20% 50px / 25% 70px)'},
  {at: 0.5, expect: 'shape(from 10% 10px, curve to 15% 5px with 5% 70px, curve to 25% 25px with 25% 45px / 10% 85px)'},
  {at: 1, expect: 'shape(from 15% 15px, curve to 20% 0px with 10% 60px, curve to 20% 30px with 30% 40px / -5% 100px)'},
  {at: 1.5, expect: 'shape(from 20% 20px, curve to 25% -5px with 15% 50px, curve to 15% 35px with 35% 35px / -20% 115px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, curve by 10% 10px with 0% 80px, curve by 30% 20px with 20% 50px / 25% 70px)',
  to: 'shape(from 15% 15px, curve by 20% 0px with 10% 60px, curve by 20% 30px with 30% 40px / -5% 100px)',
}, [
  {at: -0.3, expect: 'shape(from 2% 2px, curve by 7% 13px with -3% 86px, curve by 33% 17px with 17% 53px / 34% 61px)'},
  {at: 0, expect: 'shape(from 5% 5px, curve by 10% 10px with 0% 80px, curve by 30% 20px with 20% 50px / 25% 70px)'},
  {at: 0.5, expect: 'shape(from 10% 10px, curve by 15% 5px with 5% 70px, curve by 25% 25px with 25% 45px / 10% 85px)'},
  {at: 1, expect: 'shape(from 15% 15px, curve by 20% 0px with 10% 60px, curve by 20% 30px with 30% 40px / -5% 100px)'},
  {at: 1.5, expect: 'shape(from 20% 20px, curve by 25% -5px with 15% 50px, curve by 15% 35px with 35% 35px / -20% 115px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, smooth to 10% 10px with 0% 80px, smooth to 30% 20px)',
  to: 'shape(from 15% 15px, smooth to 20% 0px with 10% 60px, smooth to 20% 30px)',
}, [
  {at: -0.3, expect: 'shape(from 2% 2px, smooth to 7% 13px with -3% 86px, smooth to 33% 17px)'},
  {at: 0, expect: 'shape(from 5% 5px, smooth to 10% 10px with 0% 80px, smooth to 30% 20px)'},
  {at: 0.5, expect: 'shape(from 10% 10px, smooth to 15% 5px with 5% 70px, smooth to 25% 25px)'},
  {at: 1, expect: 'shape(from 15% 15px, smooth to 20% 0px with 10% 60px, smooth to 20% 30px)'},
  {at: 1.5, expect: 'shape(from 20% 20px, smooth to 25% -5px with 15% 50px, smooth to 15% 35px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, smooth by 10% 10px with 0% 80px, smooth by 30% 20px)',
  to: 'shape(from 15% 15px, smooth by 20% 0px with 10% 60px, smooth by 20% 30px)',
}, [
  {at: -0.3, expect: 'shape(from 2% 2px, smooth by 7% 13px with -3% 86px, smooth by 33% 17px)'},
  {at: 0, expect: 'shape(from 5% 5px, smooth by 10% 10px with 0% 80px, smooth by 30% 20px)'},
  {at: 0.5, expect: 'shape(from 10% 10px, smooth by 15% 5px with 5% 70px, smooth by 25% 25px)'},
  {at: 1, expect: 'shape(from 15% 15px, smooth by 20% 0px with 10% 60px, smooth by 20% 30px)'},
  {at: 1.5, expect: 'shape(from 20% 20px, smooth by 25% -5px with 15% 50px, smooth by 15% 35px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, arc to 15% -15px of 10px 20px, arc by 15% -5px of 30px cw rotate 30deg large, arc to 25% 20px of 10px 5px small)',
  to: 'shape(from 15% 15px, arc to 5% -25px of 20px 30px, arc by 25% -15px of 20px cw rotate 270deg small, arc to 25% 20px of 10px 5px small cw)'
}, [
  {at: -0.3, expect: 'shape(from 2% 2px, arc to 18% -12px of 7px 17px ccw small, arc by 12% -2px of 33px 33px rotate -42deg cw large , arc to 25% 20px of 10px 5px ccw small)'},
  {at: 0, expect: 'shape(from 5% 5px, arc to 15% -15px of 10px 20px, arc by 15% -5px of 30px cw rotate 30deg large, arc to 25% 20px of 10px 5px small)'},
  {at: 0.3, expect: 'shape(from 8% 8px, arc to 12% -18px of 13px 23px ccw small, arc by 18% -8px of 27px 27px rotate 102deg cw large, arc to 25% 20px of 10px 5px cw small )'},
  {at: 0.5, expect: 'shape(from 10% 10px, arc to 10% -20px of 15px 25px ccw small, arc by 20% -10px of 25px rotate 150deg cw large, arc to 25% 20px of 10px 5px cw small)'},
  {at: 1, expect: 'shape(from 15% 15px, arc to 5% -25px of 20px 30px, arc by 25% -15px of 20px rotate 270deg cw small, arc to 25% 20px of 10px 5px cw small)'},
  {at: 1.5, expect: 'shape(from 20% 20px, arc to 0% -30px of 25px 35px ccw small, arc by 30% -20px of 15px rotate 390deg cw small, arc to 25% 20px of 10px 5px cw small)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5px -5%, hline to 10px, vline by 10rem, hline by 8.25px, close, vline by 3pt)',
  to: 'shape(from -5px 5px, hline to 20px, vline by 10%, hline by 1em, close, vline by 6pt)',
}, [
  {at: -0.3, expect: 'shape(from 8px calc(-6.5% - 1.5px), hline to 7px, vline by calc(-3% + 208px), hline by 5.92px, close, vline by 2.8px)'},
  {at: 0, expect: 'shape(from 5px -5%, hline to 10px, vline by calc(0% + 160px), hline by 8.25px, close, vline by 4px)'},
  {at: 0.6, expect: 'shape(from -1px calc(-2% + 3px), hline to 16px, vline by calc(6% + 64px), hline by 12.9px, close ,vline by 6.4px)'},
  {at: 1, expect: 'shape(from -5px calc(0% + 5px), hline to 20px, vline by 10%, hline by 16px, close, vline by 8px)'},
  {at: 1.5, expect: 'shape(from -10px calc(2.5% + 7.5px), hline to 25px, vline by calc(15% - 80px), hline by 19.88px, close, vline by 10px)'},
]);

test_no_interpolation({
  property: 'clip-path',
  from: 'shape(from 10px 10px, move to 10% 10%)',
  to: 'path("M10 10 z")',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'path("M10 10 M10 10")',
  to: 'shape(from 10px 10px, close)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'path("M10 10 h 5")',
  to: 'shape(from 10px 10px, hline to 5px)',
});

test_no_interpolation({
  property: 'clip-path',
  from: 'shape(nonzero from 10px 10px, move to 10% 10%)',
  to: 'path(evenodd, "M0 0 M20 20")',
});

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5px 5px, hline to 5px, vline to -5px, close)',
  to: 'path("M 15 15 H 25 V -15 Z")',
}, [
  {at: -0.3, expect: 'shape(from 2px 2px, hline to -1px, vline to -2px, close)'},
  {at: 0, expect: 'shape(from 5px 5px, hline to 5px, vline to -5px, close)'},
  {at: 0.5, expect: 'shape(from 10px 10px, hline to 15px, vline to -10px, close)'},
  {at: 1, expect: 'shape(from 15px 15px, hline to 25px, vline to -15px, close)'},
  {at: 1.5, expect: 'shape(from 20px 20px, hline to 35px, vline to -20px, close)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, curve to 10% 10px with 0% 80px, curve to 30% 20px with 20% 50px / 25% 70px)',
  to: 'path("M 15 15 Q 10 60 20 0 C 30 40 -5 100 20 30")',
}, [
  {at: -0.3, expect: 'shape(from calc(6.5% - 4.5px) 2px, curve to calc(13% - 6px) 13px with calc(0% - 3px) 86px, curve to calc(39% - 6px) 17px with calc(26% - 9px) 53px / calc(32.5% + 1.5px) 61px)'},
  {at: 0, expect: 'shape(from 5% 5px, curve to 10% 10px with 0% 80px, curve to 30% 20px with 20% 50px / 25% 70px)'},
  {at: 0.5, expect: 'shape(from calc(2.5% + 7.5px) 10px, curve to calc(5% + 10px) 5px with calc(0% + 5px) 70px, curve to calc(15% + 10px) 25px with calc(10% + 15px) 45px / calc(12.5% - 2.5px) 85px)'},
  {at: 1, expect: 'shape(from calc(0% + 15px) 15px, curve to calc(0% + 20px) 0px with calc(0% + 10px) 60px, curve to calc(0% + 20px) 30px with calc(0% + 30px) 40px / calc(0% - 5px) 100px)'},
  {at: 1.5, expect: 'shape(from calc(-2.5% + 22.5px) 20px, curve to calc(-5% + 30px) -5px with calc(0% + 15px) 50px, curve to calc(-15% + 30px) 35px with calc(-10% + 45px) 35px / calc(-12.5% - 7.5px) 115px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'path("M 5 5 q 0 80 10 10 c 20 50 25 70 30 20")',
  to: 'shape(from 15% 15px, curve by 20% 0px with 10% 60px, curve by 20% 30px with 30% 40px / -5% 100px)',
}, [
  {at: -0.3, expect: 'shape(from calc(-4.5% + 6.5px) 2px, curve by calc(-6% + 13px) 13px with -3% 86px, curve by calc(-6% + 39px) 17px with calc(-9% + 26px) 53px / calc(1.5% + 32.5px) 61px)'},
  {at: 0, expect: 'shape(from calc(0% + 5px) 5px, curve by calc(0% + 10px) 10px with 0% 80px, curve by calc(0% + 30px) 20px with calc(0% + 20px) 50px / calc(0% + 25px) 70px)'},
  {at: 0.5, expect: 'shape(from calc(7.5% + 2.5px) 10px, curve by calc(10% + 5px) 5px with 5% 70px, curve by calc(10% + 15px) 25px with calc(15% + 10px) 45px / calc(-2.5% + 12.5px) 85px)'},
  {at: 1, expect: 'shape(from 15% 15px, curve by 20% 0px with 10% 60px, curve by 20% 30px with 30% 40px / -5% 100px)'},
  {at: 1.5, expect: 'shape(from calc(22.5% - 2.5px) 20px, curve by calc(30% - 5px) -5px with 15% 50px, curve by calc(30% - 15px) 35px with calc(45% - 10px) 35px / calc(-7.5% - 12.5px) 115px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, smooth to 10% 10px with 0% 80px, smooth to 30% 20px)',
  to: 'path("M 15 15 S 10 60 20 0 T 20 30")',
}, [
  {at: -0.3, expect: 'shape(from calc(6.5% - 4.5px) 2px, smooth to calc(13% - 6px) 13px with calc(0% - 3px) 86px, smooth to calc(39% - 6px) 17px)'},
  {at: 0, expect: 'shape(from 5% 5px, smooth to 10% 10px with 0% 80px, smooth to 30% 20px)'},
  {at: 0.5, expect: 'shape(from calc(2.5% + 7.5px) 10px, smooth to calc(5% + 10px) 5px with calc(0% + 5px) 70px, smooth to calc(15% + 10px) 25px)'},
  {at: 1, expect: 'shape(from calc(0% + 15px) 15px, smooth to calc(0% + 20px) 0px with calc(0% + 10px) 60px, smooth to calc(0% + 20px) 30px)'},
  {at: 1.5, expect: 'shape(from calc(-2.5% + 22.5px) 20px, smooth to calc(-5% + 30px) -5px with calc(0% + 15px) 50px, smooth to calc(-15% + 30px) 35px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'path("M 5 5 s 0 80 10 10 t 30 20")',
  to: 'shape(from 15px 15px, smooth by 20px 0px with 10px 60px, smooth by 20px 30px)',
}, [
  {at: -0.3, expect: 'shape(from 2px 2px, smooth by 7px 13px with -3px 86px, smooth by 33px 17px)'},
  {at: 0, expect: 'shape(from 5px 5px, smooth by 10px 10px with 0px 80px, smooth by 30px 20px)'},
  {at: 0.5, expect: 'shape(from 10px 10px, smooth by 15px 5px with 5px 70px, smooth by 25px 25px)'},
  {at: 1, expect: 'shape(from 15px 15px, smooth by 20px 0px with 10px 60px, smooth by 20px 30px)'},
  {at: 1.5, expect: 'shape(from 20px 20px, smooth by 25px -5px with 15px 50px, smooth by 15px 35px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'shape(from 5% 5px, arc to 15% -15px of 10px 20px, arc by 15% -5px of 30px cw rotate 30deg large, arc to 25% 20px of 10px 5px small)',
  to: 'path("M 15 15 A 20,30 0 0,0 5,-25 a 20,20 270 0,1 25,-15 A 10,5 0 0,0 25 20")',
}, [
  {at: -0.3, expect: 'shape(from calc(6.5% - 4.5px) 2px, arc to calc(19.5% - 1.5px) -12px of 7px 17px, arc by calc(19.5% - 7.5px) -2px of 33px cw large rotate -42deg, arc to calc(32.5% - 7.5px) 20px of 10px 5px)'},
  {at: 0, expect: 'shape(from 5% 5px, arc to 15% -15px of 10px 20px, arc by 15% -5px of 30px cw rotate 30deg large, arc to 25% 20px of 10px 5px small)'},
  {at: 0.3, expect: 'shape(from calc(3.5% + 4.5px) 8px, arc to calc(10.5% + 1.5px) -18px of 13px 23px, arc by calc(10.5% + 7.5px) -8px of 27px cw large rotate 102deg, arc to calc(17.5% + 7.5px) 20px of 10px 5px)'},
  {at: 0.5, expect: 'shape(from calc(2.5% + 7.5px) 10px, arc to calc(7.5% + 2.5px) -20px of 15px 25px, arc by calc(7.5% + 12.5px) -10px of 25px cw large rotate 150deg, arc to calc(12.5% + 12.5px) 20px of 10px 5px)'},
  {at: 1, expect: 'shape(from calc(0% + 15px) 15px, arc to calc(0% + 5px) -25px of 20px 30px, arc by calc(0% + 25px) -15px of 20px cw rotate 270deg, arc to calc(0% + 25px) 20px of 10px 5px)'},
  {at: 1.5, expect: 'shape(from calc(-2.5% + 22.5px) 20px, arc to calc(-7.5% + 7.5px) -30px of 25px 35px, arc by calc(-7.5% + 37.5px) -20px of 15px cw rotate 390deg, arc to calc(-12.5% + 37.5px) 20px of 10px 5px)'},
]);

test_interpolation({
  property: 'clip-path',
  from: 'path("M 5 5 A 10,20 0 0,0 15,-15 a 30,30 30 1,1 15,-5 A 10,5 0 0,0 25 20")',
  to: 'shape(from 15px 15px, arc to 5px -25px of 20px 30px, arc by 25px -15px of 20px cw rotate 270deg small, arc to 25px 20px of 10px 5px small cw)' // ccw, cw, cw
}, [
  {at: -0.3, expect: 'shape(from 2px 2px, arc to 18px -12px of 7px 17px ccw small, arc by 12px -2px of 33px 33px rotate -42deg cw large, arc to 25px 20px of 10px 5px ccw small)'},
  {at: 0, expect: 'shape(from 5px 5px, arc to 15px -15px of 10px 20px, arc by 15px -5px of 30px cw rotate 30deg large, arc to 25px 20px of 10px 5px small)'},
  {at: 0.3, expect: 'shape(from 8px 8px, arc to 12px -18px of 13px 23px ccw small, arc by 18px -8px of 27px 27px rotate 102deg cw large, arc to 25px 20px of 10px 5px cw small )'},
  {at: 0.5, expect: 'shape(from 10px 10px, arc to 10px -20px of 15px 25px ccw small, arc by 20px -10px of 25px rotate 150deg cw large, arc to 25px 20px of 10px 5px cw small)'},
  {at: 1, expect: 'shape(from 15px 15px, arc to 5px -25px of 20px 30px, arc by 25px -15px of 20px rotate 270deg cw small, arc to 25px 20px of 10px 5px cw small)'},
  {at: 1.5, expect: 'shape(from 20px 20px, arc to 0px -30px of 25px 35px ccw small, arc by 30px -20px of 15px rotate 390deg cw small, arc to 25px 20px of 10px 5px cw small)'},
]);

</script>
</body>
